backend/requirements.txt
# Django framework latest version 11.2024
Django==5.1.3
# DRF framework latest version 11.2024
djangorestframework==3.15.2
# REST authentification
dj-rest-auth==4.0.1
# implement JWT (JSON web Token)
djangorestframework-simplejwt==5.3.1
# Python library for yasg
setuptools==75.6.0
# Recommended DRF documentation OpenAPI 3
drf-spectacular==0.28.0
# Used for CORS (backend django - frontend Vue.js)
django-cors-headers==4.6.0
# django shell utilities
django-extensions==3.2.3
# Using mySQL database with Python
mysqlclient==2.2.6
# HTTP request library
requests==2.32.3
# Using python meteo france
meteofrance-api==1.3.0

